protected route functionality

+++ workshop/handlers/dogs.js
@@ -20,19 +20,34 @@ function get(req, res, next) {
 }
 
 function post(req, res, next) {
-  const newDoggy = req.body;
+  const user = req.user;
+  const dog = req.body;
+  dog.owner = user.id;
   model
-    .createDog(newDoggy)
-    .then((dog) => res.status(201).send(dog))
+    .createDog(dog)
+    .then((dog) => {
+      res.status(201).send(dog);
+    })
     .catch(next);
 }
 
 function del(req, res, next) {
-  const dogId = req.params.id
+  const dogId = req.params.id;
+  const user = req.user.id;
   model
-    .deleteDog(dogId)
-    .then(() => res.status(204).send())
-    .catch(next)
+    .getDog(dogId)
+    .then((dog)=>{
+      if (dog.owner == user) {
+        model
+          .deleteDog(dogId)
+          .then(() => res.status(204).send())
+          .catch(next)
+      } else {
+        const error = new Error("bad owner")
+        error.status = 401;
+        next(error)
+      }
+    })
 }
 
 module.exports = { getAll, get, post, del };
